Revert "Audio Added"
This reverts commit 8702add5ddda7a286ef2a0076d1353cbf4358a0d.

+++ 2D Movement Template/2D Player Movement template/Assets/Scripts/PlayerMove.cs
@@ -14,7 +14,7 @@ public class PlayerMove : MonoBehaviour
     private int spriteVersion = 0;
     public Sprite[] sprites;
     private SpriteRenderer sr;
-    public float animSpeed;
+    public int animSpeed;
 
     
     void Start()
